Refactor devtools_extensions example
