+++ packages/devtools_extensions/example/foo/packages/foo_devtools_extension/pubspec.yaml
@@ -14,6 +14,9 @@ dependencies:
   flutter:
     sdk: flutter
 
+flutter:
+  uses-material-design: true
+
 dependency_overrides:
   devtools_app_shared:
     path: ../../../../../devtools_app_shared
